Aggressively simplify pat-upload.
The original goal was to enable uploading multiple files simultaneously.

However, this pattern makes too many assumptions and is too much tied to
integration with Plone. Uploading of multiple files at once was disabled and
hardcoded. Additionally, work that could be offloaded to Dropzone was being
done in this pattern, presumably to avoid having to do any changes in the Plone
backend.

Patternslib patterns are much more generic and also don't try to serve every
single conceivable usecase, unless and until it's clear that a particular
usecase is needed.

I've therefore removed many of the options and code and tried to simplify.

Uploading multiple files simultaneously is now possible.